feat: 재고 서비스 모듈 연결

+++ module_order_service/src/main/java/com/example/module_order_service/order/domain/OrderCreate.java
@@ -9,13 +9,13 @@ public class OrderCreate {
 
     private Long productId;
     private Long memberId;
-    private Long quantity;
+    private Integer quantity;
     private String address;
 
     public OrderCreate(
             final Long productId,
             final Long memberId,
-            final Long quantity,
+            final Integer quantity,
             final String address
     ) {
         this.productId = productId;
