+++ module_order_service/src/main/java/com/example/module_order_service/order/domain/OrderProduct.java
@@ -8,10 +8,10 @@ public class OrderProduct {
 
     private String name;
     private String content;
-    private Long price;
+    private Integer price;
     private LocalDateTime purchaseButtonActivationAt;
 
-    public OrderProduct(final String name, final String content, final Long price, final LocalDateTime purchaseButtonActivationAt) {
+    public OrderProduct(final String name, final String content, final Integer price, final LocalDateTime purchaseButtonActivationAt) {
         this.name = name;
         this.content = content;
         this.price = price;
